% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashprng-package.R
\docType{package}
\name{hashprng}
\alias{hashprng}
\title{hashprng: Hash-Based Matching Pseudo-Random Number Generation}
\description{
The hashprng package provides single function for use during stochastic
simulation to streamline salting + event hashing + reseeding the R random
number generator. The sole function in the package is \code{hash_seed}.
}
