% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbmPRNG-package.R, R/hbmPRNG.R
\docType{package}
\name{hbmPRNG}
\alias{hbmPRNG}
\title{hbmPRNG: Hash-Based Matching Pseudo-Random Number Generation}
\description{
The hbmPRNG package provides single function for use during stochastic
simulation to streamline salting + event hashing + reseeding the R random
number generator. The sole function in the package is \code{hash_seed}.

The hbmPRNG package provides single function for use during stochastic
simulation to streamline salting + event hashing + reseeding the R random
number generator. The sole function in the package is \code{hash_seed}.
}
