% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{r2dtable}
\alias{r2dtable}
\title{Hash-Matched r2dtable}
\usage{
r2dtable(n, r, c, hash)
}
\arguments{
\item{n}{a non-negative numeric giving the number of tables to be
    drawn.}

\item{r}{a non-negative vector of length at least 2 giving the row
    totals, to be coerced to \code{integer}.  Must sum to the same as
    \code{c}.}

\item{c}{a non-negative vector of length at least 2 giving the column
    totals, to be coerced to \code{integer}.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched r2dtable
}
