% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rbeta}
\alias{rbeta}
\title{Hash-Matched rbeta}
\usage{
rbeta(n, shape1, shape2, ncp = 0, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{shape1, shape2}{non-negative parameters of the Beta distribution.}

\item{ncp}{non-centrality parameter.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rbeta
}
