% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rbinom}
\alias{rbinom}
\title{Hash-Matched rbinom}
\usage{
rbinom(n, size, prob, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{size}{number of trials (zero or more).}

\item{prob}{probability of success on each trial.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rbinom
}
