% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rchisq}
\alias{rchisq}
\title{Hash-Matched rchisq}
\usage{
rchisq(n, df, ncp = 0, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{df}{degrees of freedom (non-negative, but can be non-integer).}

\item{ncp}{non-centrality parameter (non-negative).}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rchisq
}
