% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rexp}
\alias{rexp}
\title{Hash-Matched rexp}
\usage{
rexp(n, rate = 1, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{rate}{vector of rates.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rexp
}
