% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rf}
\alias{rf}
\title{Hash-Matched rf}
\usage{
rf(n, df1, df2, ncp, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{df1, df2}{degrees of freedom.  \code{Inf} is allowed.}

\item{ncp}{non-centrality parameter. If omitted the central F is assumed.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rf
}
