% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rgamma}
\alias{rgamma}
\title{Hash-Matched rgamma}
\usage{
rgamma(n, shape, rate = 1, scale = 1/rate, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{shape, scale}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{rate}{an alternative way to specify the scale.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rgamma
}
