% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rgeom}
\alias{rgeom}
\title{Hash-Matched rgeom}
\usage{
rgeom(n, prob, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rgeom
}
