% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rhyper}
\alias{rhyper}
\title{Hash-Matched rhyper}
\usage{
rhyper(nn, m, n, k, hash)
}
\arguments{
\item{nn}{number of observations.  If \code{length(nn) > 1}, the length
    is taken to be the number required.}

\item{m}{the number of white balls in the urn.}

\item{n}{the number of black balls in the urn.}

\item{k}{the number of balls drawn from the urn, hence must be in
    \eqn{0,1,\dots, m+n}.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rhyper
}
