% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rlnorm}
\alias{rlnorm}
\title{Hash-Matched rlnorm}
\usage{
rlnorm(n, meanlog = 0, sdlog = 1, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{meanlog, sdlog}{mean and standard deviation of the distribution
    on the log scale with default values of \code{0} and \code{1} respectively.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rlnorm
}
