% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rlogis}
\alias{rlogis}
\title{Hash-Matched rlogis}
\usage{
rlogis(n, location = 0, scale = 1, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{location, scale}{location and scale parameters.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rlogis
}
