% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rmultinom}
\alias{rmultinom}
\title{Hash-Matched rmultinom}
\usage{
rmultinom(n, size, prob, hash)
}
\arguments{
\item{n}{number of random vectors to draw.}

\item{size}{integer, say \eqn{N}, specifying the total number
    of objects that are put into \eqn{K} boxes in the typical multinomial
    experiment. For \code{dmultinom}, it defaults to \code{sum(x)}.}

\item{prob}{numeric non-negative vector of length \eqn{K}, specifying
    the probability for the \eqn{K} classes; is internally normalized to
    sum 1. Infinite and missing values are not allowed.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rmultinom
}
