% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rnbinom}
\alias{rnbinom}
\title{Hash-Matched rnbinom}
\usage{
rnbinom(n, size, prob, mu, hash)
}
\arguments{
\item{n}{number of observations.  If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{size}{target for number of successful trials, or dispersion
    parameter (the shape parameter of the gamma mixing distribution).
    Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}

\item{mu}{alternative parametrization via mean: see \sQuote{Details}.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rnbinom
}
