% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rnorm}
\alias{rnorm}
\title{Hash-Matched rnorm}
\usage{
rnorm(n, mean = 0, sd = 1, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rnorm
}
