% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rpois}
\alias{rpois}
\title{Hash-Matched rpois}
\usage{
rpois(n, lambda, hash)
}
\arguments{
\item{n}{number of random values to return.}

\item{lambda}{vector of (non-negative) means.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rpois
}
