% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrejig.R
\name{rrejig}
\alias{rrejig}
\title{Add Hash-seeding to a Deviate Generators}
\usage{
rrejig(rFUN)
}
\arguments{
\item{rFUN}{a function}
}
\value{
a new function, which should have all the same arguments
plus a new \code{hash} argument at the end.
}
\description{
This function provides a way extend the arguments of deviate
generating functions -- that is, the \code{runif}, \code{rbinom}, etc.

Internally, \code{hashprng} applies this to commonly used \code{{stats}}
functions. When you attach \code{hashprng}, those are all masked.
}
