% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rsignrank}
\alias{rsignrank}
\title{Hash-Matched rsignrank}
\usage{
rsignrank(nn, n, hash)
}
\arguments{
\item{nn}{number of observations. If \code{length(nn) > 1}, the length
    is taken to be the number required.}

\item{n}{number(s) of observations in the sample(s).  A positive
    integer, or a vector of such integers.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rsignrank
}
