% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rsmirnov}
\alias{rsmirnov}
\title{Hash-Matched rsmirnov}
\usage{
rsmirnov(n, sizes, z = NULL, two.sided = TRUE, hash)
}
\arguments{
\item{n}{an integer giving number of observations.}

\item{sizes}{an integer vector of length two giving the sample sizes.}

\item{z}{a numeric vector of the pooled data values in both samples
    when the exact conditional distribution of the Smirnov statistic
    given the data shall be computed.}

\item{two.sided}{a logical indicating whether absolute (\code{TRUE}) or
    raw differences of frequencies define the test statistic.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rsmirnov
}
