% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rstudent}
\alias{rstudent}
\title{Hash-Matched rstudent}
\usage{
rstudent(model, ..., hash)
}
\arguments{
\item{model}{an \R object, typically returned by \code{\link[stats]{lm}} or
    \code{\link[stats]{glm}}.}

\item{...}{further arguments passed to or from other methods.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rstudent
}
