% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rt}
\alias{rt}
\title{Hash-Matched rt}
\usage{
rt(n, df, ncp, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{df}{degrees of freedom (\eqn{> 0}, maybe non-integer).  \code{df
      = Inf} is allowed.}

\item{ncp}{non-centrality parameter \eqn{\delta}{delta};
    currently except for \code{rt()}, only for \code{abs(ncp) <= 37.62}.
    If omitted, use the central t distribution.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rt
}
