% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{runif}
\alias{runif}
\title{Hash-Matched runif}
\usage{
runif(n, min = 0, max = 1, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{min, max}{lower and upper limits of the distribution.  Must be finite.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched runif
}
