% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rweibull}
\alias{rweibull}
\title{Hash-Matched rweibull}
\usage{
rweibull(n, shape, scale = 1, hash)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{shape, scale}{shape and scale parameters, the latter defaulting to 1.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rweibull
}
