% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rwilcox}
\alias{rwilcox}
\title{Hash-Matched rwilcox}
\usage{
rwilcox(nn, m, n, hash)
}
\arguments{
\item{nn}{number of observations. If \code{length(nn) > 1}, the length
    is taken to be the number required.}

\item{m, n}{numbers of observations in the first and second sample,
    respectively.  Can be vectors of positive integers.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rwilcox
}
