% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_distros.R
\name{rWishart}
\alias{rWishart}
\title{Hash-Matched rWishart}
\usage{
rWishart(n, df, Sigma, hash)
}
\arguments{
\item{n}{integer sample size.}

\item{df}{numeric parameter, \dQuote{degrees of freedom}.}

\item{Sigma}{positive definite (\eqn{p\times p}{p * p}) \dQuote{scale}
    matrix, the matrix parameter of the distribution.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched rWishart
}
