% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrejig.R
\name{sample.int}
\alias{sample.int}
\title{Hash-Matched sample.int}
\usage{
sample.int(
  n,
  size = n,
  replace = FALSE,
  prob = NULL,
  useHash = (n > 1e+07 && !replace && is.null(prob) && size <= n/2),
  hash
)
}
\arguments{
\item{n}{a positive number, the number of items to choose from.  See
    \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}

\item{useHash}{\code{\link[base]{logical}} indicating if the hash-version of
    the algorithm should be used.  Can only be used for \code{replace =
      FALSE}, \code{prob = NULL}, and \code{size <= n/2}, and really
    should be used for large \code{n}, as \code{useHash=FALSE} will use
    memory proportional to \code{n}.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched sample.int
}
