% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrejig.R
\name{sample}
\alias{sample}
\title{Hash-Matched sample}
\usage{
sample(x, size, replace = FALSE, prob = NULL, hash)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose,
    or a positive integer.  See \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}

\item{hash}{an object identifying the event related to the draw}
}
\description{
Hash-Matched sample
}
