% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_salt.R
\name{set.hash}
\alias{set.hash}
\title{Manage RNG state}
\usage{
set.hash(hash, .hash.salt)
}
\arguments{
\item{hash}{an object which identifies the event}

\item{.hash.salt}{an object which identifies the universe}
}
\description{
Manage RNG state
}
