% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_salt.R
\name{set.salt}
\alias{set.salt}
\title{Set Hash Salt}
\usage{
set.salt(salt)
}
\arguments{
\item{salt}{an object that will be used as the starter object for
all hash-matched. A plain integer is sufficient.}
}
\description{
This function sets a environment value to be used for hash-matched
pseudo-random number generation (PRNG). This means that \code{salt}
values are scoped.

This is the conceptual equivalent of \code{\link[=set.seed]{set.seed()}} with unmatched
PRNG.
}
