% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashprng-package.R
\name{hashprng}
\alias{hashprng}
\title{hashprng: Hash-matched Pseudo-Random Number Generation}
\description{
The \code{{hashprng}} package simplifies using hash-matched pseudo-random
number generation.
}
